pyproject.toml
[tool.poetry]
name = "python-project-templates"
version = "0.1.0"
description = "Python poetry project templates"
authors = ["Jasper Ginn <jasperginn@gmail.com>"]
maintainers = ["Jasper Ginn <jasperginn@gmail.com>"]
repository = "https://github.com/JasperHG90/python-project-templates"

[tool.poetry.dependencies]
python = "^3.8, <3.10"
cookiecutter = "1.7.2"
toml = "0.10.2"
PyYAML = "5.4.1"
mkdocs = "1.2.1"
mkdocs-cinder = "1.2.0"

[tool.poetry.dev-dependencies]
black = "20.8b1"
pre-commit = "2.12.1"
pytest = "5.2"
pytest-cov = "2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
